pyproject.toml
[tool.poetry]
name = "pdf-page-extractor"
version = "0.1.0"
description = "Extract selected pages from PDF files and optionally prepend a Markdown introduction"
authors = ["Robert Kraig <robertkraig@gmail.com>"]
readme = "README.md"

[[tool.poetry.packages]]
include = "pdf_extractor"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pypdf2 = "^3.0.1"
pyyaml = "^6.0"
reportlab = "^4.0.0"
markdown2 = "^2.4.13"
weasyprint = "^61.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
